+++ source/Core/Utils.php
@@ -1073,8 +1073,8 @@ class Utils extends \OxidEsales\Eshop\Core\Base
 
         try { //may occur in case db is lost
             $this->getSession()->freeze();
-        } catch (\OxidEsales\Eshop\Core\Exception\StandardException $oEx) {
-            $oEx->debugOut();
+        } catch (\OxidEsales\Eshop\Core\Exception\StandardException $exception) {
+            \OxidEsales\Eshop\Core\Registry::getLogger()->error($exception->getMessage(), [$exception]);
             //do nothing else to make sure the redirect takes place
         }
 
@@ -1301,8 +1301,9 @@ class Utils extends \OxidEsales\Eshop\Core\Base
             if (gettype($sText) != 'string') {
                 $sText = var_export($sText, true);
             }
-            $sLogMsg = "----------------------------------------------\n{$sText}" . (($blNewline) ? "\n" : "") . "\n";
-            $this->writeToLog($sLogMsg, "log.txt");
+            $logMessage = "----------------------------------------------\n{$sText}" . (($blNewline) ? "\n" : "") . "\n";
+            $logger = Registry::getLogger();
+            $logger->debug($logMessage);
         }
     }
 
@@ -1386,29 +1387,6 @@ class Utils extends \OxidEsales\Eshop\Core\Base
         }
 
         return $sUrl;
-    }
-
-    /**
-     * Writes given log message. Returns write state
-     *
-     * @deprecated since v5.3 (2016-06-17); Logging mechanism will change in the future.
-     *
-     * @param string $logMessage  log message
-     * @param string $logFileName log file name
-     *
-     * @return bool
-     */
-    public function writeToLog($logMessage, $logFileName)
-    {
-        $logFilePath = $this->getConfig()->getLogsDir() . $logFileName;
-        $writeSucceed = false;
-
-        if (($logFileResource = fopen($logFilePath, 'a')) !== false) {
-            fwrite($logFileResource, $logMessage);
-            $writeSucceed = fclose($logFileResource);
-        }
-
-        return $writeSucceed;
     }
 
     /**